+++ lib/elastomer/client/docs.rb
@@ -257,22 +257,44 @@ module Elastomer
         response.body
       end
 
-      # Delete documents from one or more indices and one or more types based
+      # Delete documents by query following either the native or
+      # application-level delete by query method.
+      #
+      # NOTE: The parameters and response format varies by version. To have more
+      # control over this, use app_delete_by_query or native_delete_by_query
+      # directly.
+      def delete_by_query(query, params = nil)
+        send(@client.version_support.delete_by_query_method, query, params)
+      end
+
+      # DEPRECATED: Delete documents from one or more indices and one or more types based
       # on a query. This method supports both the "request body" query and the
       # "URI request" query. When using the request body semantics, the query
       # hash must contain the :query key. Otherwise we assume a URI request is
       # being made.
       #
-      # See Client#delete_by_query for more information.
+      # See Client#app_delete_by_query for more information.
       #
       # Returns a Hash of statistics about the delete operations
-      def delete_by_query(query, params = nil)
+      def app_delete_by_query(query, params = nil)
         query, params = extract_params(query) if params.nil?
 
-        @client.delete_by_query(query, update_params(params))
+        @client.app_delete_by_query(query, update_params(params))
       end
 
+      # Delete documents from one or more indices and one or more types based
+      # on a query using Elasticsearch's _delete_by_query API.
+      #
+      # See Client#native_delete_by_query for more information.
+      #
+      # Returns a Hash of statistics about the delete operations as returned by
+      # _delete_by_query.
+      #
+      # Raises Elastomer::Client::IncompatibleVersionException if this version
+      # of Elasticsearch does not support _delete_by_query.
       def native_delete_by_query(query, params = {})
+        query, params = extract_params(query) if params.nil?
+
         @client.native_delete_by_query(query, update_params(params))
       end
