+++ test/client/index_test.rb
@@ -330,20 +330,25 @@ describe Elastomer::Client::Index do
       @index.docs("foo").index("foo" => "bar")
       @index.refresh
       r = @index.delete_by_query(:q => "*")
-      assert_equal({
-        "_all" => {
-          "found" => 1,
-          "deleted" => 1,
-          "missing" => 0,
-          "failed" => 0,
-        },
-        @name => {
-          "found" => 1,
-          "deleted" => 1,
-          "missing" => 0,
-          "failed" => 0,
-        }
-      }, r["_indices"])
+
+      if supports_native_delete_by_query?
+        assert_equal(1, r['deleted'])
+      else
+        assert_equal({
+          "_all" => {
+            "found" => 1,
+            "deleted" => 1,
+            "missing" => 0,
+            "failed" => 0,
+          },
+          @name => {
+            "found" => 1,
+            "deleted" => 1,
+            "missing" => 0,
+            "failed" => 0,
+          }
+        }, r["_indices"])
+      end
     end
 
     it "creates a Percolator" do
